Rename msw handler file more specify
